v3 of stockscraper
this version was able to download AMEX on a friday afternoon, likely because it wasn't busy.

though it is now crash-proof, it still lacks the mechanism for retrying repeatedly when encountering a 503 error. we think what it will do is skip a stock if there is a 503 rather than retry.